+++ dst/app/routes/sellers.js
@@ -36,7 +36,7 @@ sellersRouter.get('', permitScopes_1.default(['sellers.*', 'sellers.read', 'pos'
         });
         const { data, totalCount } = yield sellerService.search(Object.assign(Object.assign(Object.assign({}, req.query), { project: { id: { $eq: req.project.id } } }), (!req.isAdmin) ? { $projection: { 'paymentAccepted.gmoInfo.shopPass': 0 } } : undefined));
         if (typeof totalCount === 'number') {
-            res.set('X-Total-Count', totalCount.toString());
+            // res.set('X-Total-Count', totalCount.toString());
         }
         res.json(data);
     }