+++ src/app/routes/events.ts
@@ -98,10 +98,6 @@ eventsRouter.get(
 
             const searchEventsResult = await eventService.search(searchConditions);
 
-            if (typeof searchEventsResult.totalCount === 'number') {
-                // res.set('X-Total-Count', searchEventsResult.totalCount.toString());
-            }
-
             res.json(searchEventsResult.data);
         } catch (error) {
             next(error);