+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @cinerino/factory
 - update @pecorino/api-nodejs-client
 - ポイント口座での決済取引開始時のfromLocationのタイプをAccountに変更
+- 販売者検索からX-Total-Countを削除
 
 ### Deprecated